+++ src/main/java/com/mahdi/sandogh/model/installmentloan/dto/InstallmentLoanForm.java
@@ -6,9 +6,9 @@ import lombok.Data;
 @Data
 public class InstallmentLoanForm {
 
-    private Long loanId;  // شماره وام
+    private Integer loanId;  // شماره وام
     private Long installmentLoanId;
-    private Long amountInstallment; // مبلغ قسط
+    private Integer amountInstallment; // مبلغ قسط
     private Integer numberLoan; // شماره قسط
     private Long accountId;
